# apispec.yaml
openapi: 3.1.1
info:
  title: Kontrak Api Toko Online
  description: Dokumentasi backend e-commerce Toko-Online dengan integrasi Redis, Cloudinary, dan Midtrans.
  version: 1.2.0
servers:
  - url: http://localhost/api
    description: Production Proxy via Nginx
paths:
...
  /orders/{id}:
    get:
      operationId: getOrderByID
      tags: [Admin]
      security:
        - BearerAuth: []
      summary: Detail pesanan berdasarkan ID
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Pesanan ditemukan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
    patch:
      operationId: cancelOrder
      tags: [Admin]
      security:
        - BearerAuth: []
      summary: Batalkan pesanan dan kembalikan stok (Admin Only)
      responses:
        "200":
          description: Pesanan dibatalkan
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    CsrfToken:
      type: apiKey
      name: X-CSRF-Token
      in: header
  schemas:
    Product:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        slug: { type: string }
        description: { type: string }
        price: { type: number }
        stock: { type: integer }
        category_id: { type: string, format: uuid }
        is_active: { type: boolean }
        image_url: { type: string }
        images:
          type: array
          items: { $ref: "#/components/schemas/ProductImage" }
        category: { $ref: "#/components/schemas/Category" }
    ProductImage:
      type: object
      properties:
        id: { type: string, format: uuid }
        product_id: { type: string, format: uuid }
        url: { type: string }
    Category:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        slug: { type: string }
        image_url: { type: string }
        product_count: { type: integer }
    OrderItem:
      type: object
      properties:
        id: { type: string, format: uuid }
        order_id: { type: string, format: uuid }
        product_id: { type: string, format: uuid }
        quantity: { type: integer }
        price: { type: number }
        product: { $ref: "#/components/schemas/Product" }
    Order:
      type: object
      properties:
        id: { type: string, format: uuid }
        order_number: { type: string }
        customer_name: { type: string }
        customer_email: { type: string }
        customer_phone: { type: string }
        shipping_address: { type: string }
        total_amount: { type: number }
        status: { type: string }
        note: { type: string }
        whatsapp_url: { type: string }
        snap_token: { type: string }
        snap_redirect_url: { type: string }
        payment_type: { type: string }
        created_at: { type: string, format: date-time }
        items:
          type: array
          items: { $ref: "#/components/schemas/OrderItem" }
...